Added dynamic antenna changes for GUI control
Addedd sepaate settings for defaults and current in the antenna class so that GUI can control the states. Note the default values - they are set out of range as a default. In a future update the check for theta_c in the antenna init will be removed since the program expects the user to check for changes (see station.h, reset_gui) in state.

Also note the separate reset functions for init_gui() and reset() from the station.h file. This commit does not update the calculations with COW changes upon receiving GUI requests. Need to add that and call antenna.numerical_update