Interference function added
Now added a separate function to compute the interference repeatedly from the render and plot funcitons

The states of each transmitter can be passed into the graphics functions more efficiently.

Added some cached functions such as watt2dbm and lin2B. Added str() function for coordinates, str() for cow, and str() for antenna State.

Moves the tx_lines subroutine into a separate function for future when visualization scales up and need to move the TX upon user request.

Lot of renames, and some refactor.

This commit leaves the inefficient heat computation as is, but consider only computing the log version of the intermediate calculations if debug is true (default: should be interference mode ...)