Graphics threads - revisited and working
Note: graphics mutex on both threads inside sim and visuals side.

The mutex can be adjusted in the future. The queue is redone (emplace, pop, front), and uses interference function from the previous commit!

grid_update is replaced with more specific type of signal(s) depending on the situation. If it's a simple color update on the grid:

render_tx_id -> update

If it is a transmitter change:

compute_tx-id -> update

In the future resize will be a thing. Lots of rename. dbm to dBm in common.h (some slight tab/space issue added here)