Visuals.h overhaul
Break tx_lines into two different functions with two meaningful names. Rotation update also called the scan angle update to update the arrows

Fixed a bug where the HeatGrid was being passed a const init state-list. It should have been curr (current states) can be changed from the render loop

Use the API created in DataSync for resize. Add rotation and scan-angle updates within the ImGui controls

resize() in this file shouldn't do anything so in the future commit going to call higher levels for recompute the grid data. It is also possible that in order to resize, there will be significant overhead in data rellocations in CUDA functions in .cu file so going to leave this feature for future as needed.