Visuals.h TX moving bug resolved
Dragginga and dropping TX to another part of the grid used to be broken.

Separate the slider and input boxes for exact inputs. Input boxes now wait for the user to finish typing a new value.

There is still an issue with slider controls blasting a bunch of updates to the render queue. Without CUDA this is very laggy.

Renamed a variable inside legend, signal_value