Legend fixed in visuals.h
Fixed a range bug since the max and min color values in the slider control were not based on the actual max and min color values queried earlier.

Adjusting color range based on whether it is debug more or interferernce in the "debug" check inside the render loop.

Renamed rows and cols, to known_width and known_height in project.h