Move min and max compute_colorspan to visuals.h
Refactor. setup_cow_heat() function still in project.h and sends this data into the render() and capture() function in dBm.

This is for debug only, so in the future this compute will be delayed as much as possible when needed using the logify funciton. This will favor the interference since it does not need to convert from dBm to lin and then lin back to dB for final SINR values.

In the commits before this the syntax shown in the generate_() function did not have the first type of translation because this was missed out. But in working code this first translation should have been there to work correctly.