CUDA file for faster calculations in network.h
Antenna updates need 2 major functions init() and update(). These are slow 5-10 seconds before an update is shown on GUI. The CUDA solution is going to be speed it up using memory pinned vectors such as polar-coordinates as inputs, and hmatrix as outputs.

That way we don't have to copy the inputs and outputs into and from the __device__ or GPU

Future updates in non-CUDA solution will be influenced by the API and data structure of CUDA solution.