move intermediate calc to visuals.h
Did not convert the lin values in raw_cow for individual signal heat into dB. But instead going to wait for the user to request it when needed. Going to copy and convert this separately when running the render loop from lin to dB. Should be faster to deploy interference model now.