Add more functions for modularity
Save feature to save the plots. Will be used from the project.h level instead of calling a separate function.

Draw fucntion in visuals.h to draw and call this too from project.h level. Needed another HeatGrid constructor but can be moved outside and use the common one as a unique_ptr across the render() and plot() functions in projec.h

Lots of rename, and reorder the initialization list. Some input parameters