Debugging mode is enabled fully
Now user can choose to check the heat coming from individual transmitters as needed. This is the third 2d data array needed after translating the linear power to dBm power.

Only when toggling to the debug mode, the program will allocate and fill in the memory needed to show the values for debug (heat values). The heat-checker thread takes care of this but the thread to can be split into 2, thus simplying the logic further.

Moved the translation function from visuals.h to project.h and added an async version of the interference calculation function. This will enable computing the interference after individual transmitters have been updated in 2 steps: update the transmitters and fill in the grid linear power values, then translate the power to dBm values. Finally, async-calculated the interference values.

This won't bog down thread if the primary viewing mode is debugging. When the viewing mode is inteference:defualt; the state of the system is captured and debugging values are computed again before rendering into the grid. This is to make sure that any extra compute not needed right away is delayed until it is needed.